update search request
